Combine actions and reducers folders
