+++ src/components/Contacts/Contacts.container.js
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux';
 import Contacts from './Contacts';
-import {addContact, refreshContacts} from '../../actions/contacts';
+import {addContact, refreshContacts} from '../../state/contacts';
 
 const addNewContact = dispatch =>
   async function(contact) {
