+++ src/pages/ProfilePage/ProfilePage.container.spec.js
@@ -1,5 +1,5 @@
 import {mapStateToProps} from './ProfilePage.container';
-import {RESUME_CREATION} from '../../reducers/resume';
+import {RESUME_CREATION} from '../../state/resume';
 
 const blankState = {
   contacts: [],
