+++ src/state/resume.js
@@ -1,6 +1,7 @@
 import {format} from 'date-fns';
 import {API_URL} from '../constants';
 import {makeFetchActions, fetchActionTypes} from 'redux-fetch-wrapper';
+import {createReducer} from 'redux-starter-kit';
 
 export const START_RESUME_CREATION = 'START_RESUME_CREATION';
 export const startResumeCreation = () => ({
@@ -229,3 +230,85 @@ export const deleteResumeSection = (resumeId, sectionId) =>
       return success;
     }
   };
+
+export const RESUME_CREATION = {
+  NOT_ACTIVE: 'NOT_ACTIVE',
+  CHOOSE_STYLE: 'CHOOSE_STYLE',
+  SELECT_HIGHLIGHTS: 'SELECT_HIGHLIGHTS',
+};
+
+const getExperienceKey = experience => {
+  const expType = experience.type;
+  if (expType === 'Education') {
+    return 'education';
+  } else if (expType === 'Accomplishment') {
+    return 'accomplishments';
+  } else {
+    return 'experience';
+  }
+};
+
+const genInitState = () => ({
+  resumeCreationStep: RESUME_CREATION.NOT_ACTIVE,
+  inProgress: false,
+  selected: {
+    experience: [],
+    education: [],
+    accomplishments: [],
+  },
+  resumes: [],
+});
+
+export const resumeReducer = createReducer(genInitState(), {
+  [START_RESUME_CREATION]: (state, action) => {
+    state.resumeCreationStep = RESUME_CREATION.CHOOSE_STYLE;
+  },
+  [START_RESUME_SELECT]: (state, action) => {
+    state.resumeCreationStep = RESUME_CREATION.SELECT_HIGHLIGHTS;
+  },
+  [CANCEL_RESUME_SELECT]: (state, action) => {
+    Object.keys(state.selected).forEach(key => {
+      state.selected[key] = [];
+    });
+    state.resumeCreationStep = RESUME_CREATION.NOT_ACTIVE;
+  },
+  [SELECT_RESUME_EXPERIENCE]: (state, action) => {
+    // Only allow item selection when we're in the right state for it
+    // Hopefully this should keep our state clean across oddly timed
+    // transitions, etc.
+    if (state.resumeCreationStep !== RESUME_CREATION.SELECT_HIGHLIGHTS) {
+      return;
+    }
+
+    const key = getExperienceKey(action.experience);
+    state.selected[key].push(action.experience.id);
+  },
+
+  [DESELECT_RESUME_EXPERIENCE]: (state, action) => {
+    const key = getExperienceKey(action.experience);
+    state.selected[key] = state.selected[key].filter(
+      id => id !== action.experience.id
+    );
+  },
+  [DESELECT_RESUME_EXPERIENCE]: (state, action) => {
+    const key = getExperienceKey(action.experience);
+    state.selected[key] = state.selected[key].filter(
+      id => id !== action.experience.id
+    );
+  },
+  [GENERATE_RESUME_API.REQUEST]: (state, action) => {
+    state.inProgress = true;
+  },
+  [GENERATE_RESUME_API.RESOLVE]: (state, action) => {
+    const initState = genInitState();
+    state.resumeCreationStep = initState.resumeCreationStep;
+    state.selected = initState.selected;
+    state.resumes.push(action.body.data);
+    state.inProgress = false;
+  },
+  [GENERATE_RESUME_API.REJECT]: (state, action) => {
+    state.inProgress = false;
+  },
+});
+
+export default resumeReducer;
